refactor: put cli in cli
